+++ addOns/insights/CHANGELOG.md
@@ -4,7 +4,8 @@ All notable changes to this add-on will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## Unreleased
-
+### Fixed
+- Address concurrency issue while generating report with insights.
 
 ## [0.0.1] - 2025-12-15